+++ tests/test_training_runtime.py
@@ -335,7 +335,7 @@ def test_training_cli_defaults_to_per_device_batch_eight(
 	assert args.dataset_root == tmp_path / "coco"
 	assert args.per_device_batch_size == 8
 	assert args.attention_implementation == "auto"
-	assert args.runtime_precision == "bf16"
+	assert args.runtime_precision == "fp16"
 	assert args.initial_gradient_scale == 65_536.0
 	assert args.max_checkpoints == 4
 	assert args.training_config == Path("configs/train.yaml")
